tests/fixtures/StrayRatz/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
Flask-Migrate==4.0.5
email-validator==2.0.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
Flask-Limiter==3.12
gunicorn==21.2.0
psycopg2-binary==2.9.9